requirements.txt
pytesseract==0.2.6
face_recognition==1.2.3
numpy==1.15.4
Pillow==6.0.0
Flask==1.1.1
Flask-Compress==1.4.0
Flask-Cors==3.0.8
Flask-WTF==0.14.2
WTForms==2.2.1
termcolor==1.1.0
requests==2.23.0
requests-oauthlib==1.3.0
